pyproject.toml
[project]
name = "photo-organizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.136.1",
    "imagehash>=4.3.2",
    "jinja2>=3.1.6",
    "pillow>=12.2.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn[standard]>=0.46.0",
]
